+++ CHANGELOG.md
@@ -1,11 +1,17 @@
+## 0.1.2
+
+* add example main
+* CustomKeyBoard change CAPS, DELETE to ⇧, ←
+* CustomKeyBoard add buttonPadding, button hSize, button wSize, buttonRadius, buttonColor
+
 ## 0.1.1
 
 * add usage in README.md
-*
+
 ## 0.1.0
 
 * add usage in README.md
-* 
+
 ## 0.0.1
 
 * Initial release.